+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">travis</string>
+    <string name="app_name">Do what you want</string>
 </resources>
