+++ app/src/test/java/com/lockwood/dwyw/ExampleUnitTest.kt
@@ -1,8 +1,7 @@
-package com.lockwood.travis
+package com.lockwood.dwyw
 
+import org.junit.Assert.assertEquals
 import org.junit.Test
-
-import org.junit.Assert.*
 
 /**
  * Example local unit test, which will execute on the development machine (host).
@@ -10,8 +9,8 @@ import org.junit.Assert.*
  * See [testing documentation](http://d.android.com/tools/testing).
  */
 class ExampleUnitTest {
-    @Test
-    fun addition_isCorrect() {
-        assertEquals(4, 2 + 2)
-    }
+	@Test
+	fun addition_isCorrect() {
+		assertEquals(4, 2 + 2)
+	}
 }
